requirements.txt
openai-whisper
torch
sounddevice
scipy
soundfile
numpy
pydub
ffmpeg-python